Remove unneeded unit tests for unneeded UUID method
getMojangUUID was a replacement for the regex conversion of UUIDs to
Mojang's format. I have removed this method and the associated unit
tests because it is nothing more than a wrapper around String#replace.
This new patch simply replaces the call to String#replaceAll with a
call to String#replace. Patch 0011 was affected by this change, and has
been fixed.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>