Provide an option to disable entity metadata rewriting (fixes #136)
Closes #309

The work here is derived from the research and guidance of
various members of the minecraft community

This patch provides the ability to disable entity remapping,
which creates various incompatability issues with mods, however,
may also create various issues with mods which do not support this,
hence why the configuration option is provided.

This patchs changes will be disabled by default, as we do not
wish to release an update which may cause certain issues with mods
and plugins by default when this is not essential, however, this
is subject to be changed in the future.