Fix switching servers with disable metadata rewrite
The issue happened when you where in the nether and you changed from that server
to another server that was in the overworld. You would stay stuck endlessly.

In addition, BungeeCord changed how the dimension changed internally, so we can
safely cast to an Integer without having to worry about if the dimension was an
object or an int.