Don't send exceptions to the client during kicks, etc
The info here is often not filtered down and can often contain
internal information during these kicks which server owners are
generally not happy with sharing, i.e. potential IP addresses, etc

We'll have to dump the cause to the console manually to ensure that
data is still sent, but, a wider fix here is probably needed which
allows for retaining much of the overall context here, i.e. who
was this exception assocated with?